+++ example/a.ts
@@ -1,18 +1,30 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-new Elysia()
-	.macro({
-		token: {
-			resolve: () => {
-				return {
-					__token: '123'
-				}
+const app = new Elysia()
+	.wrap((fn, request) => {
+		const _request = request.clone()
+
+		return () => {
+			try {
+				return fn(request.clone())
+			} catch {
+				console.log('ER')
 			}
 		}
 	})
-	.macro('some', {
-		token: true,
-		beforeHandle: ({ __token }) => {
-			console.log('__token', __token)
-		}
+	.onError(({ error }) => {
+		if (error) throw error
 	})
+	.get('/', () => {
+		throw new Error('A')
+
+		return 'Hello World!'
+	})
+	.listen(3000)
+
+// console.log(app.fetch.toString())
+
+app.handle(req('/'))
+	.then((x) => x.text())
+	.then(console.log)